Moved some storyqueries from contribs. Storyview now shows up on story click.
