+++ public/index.html
@@ -92,12 +92,12 @@
     <div id="story-view" class="col-md-8 container-md">
       <section class="app-hero">
 
-        <div class="story-creation-form">
+        <!-- <div class="story-creation-form">
           <input type="text" id="new-story-title" placeholder="Enter your title.">
           <textarea id="new-story-text" placeholder="Begin your story!"></textarea>
           <button id="submit-story">CREATE</button>
           <button id="cancel-story">Cancel</button>
-        </div>
+        </div> -->
 
         <h1>Story</br><b>Creator</b></h1>
         <h3>Tell your story.</h3>
