+++ public/scripts/s-story-view.js
@@ -14,11 +14,11 @@ $(document).ready(() => {
      } = story;
 
      const $storyView= `
-      <article class="storyView" id="article${id}">
+      <article class="story-view" id="${id}">
 
-      <header class="storyViewHeader">
+      <header class="story-view-header">
       <h4>${title}</h4>
-      <i>Creator: #${user_id}</i>
+      <i>Created By: #${user_id}</i>
       </header>
 
       <p>${main_story}</p>
@@ -37,20 +37,16 @@ $(document).ready(() => {
   $("#story-list").on("click", 'article[id]', function () {
     // 'this' refers to the clicked div
     var storyId = $(this).attr("id");
-    alert("Clicked on story#" + storyId); //test
+    // alert("Clicked on story#" + storyId); //test
 
     $.ajax({
       method: "GET",
       url: `/stories/${storyId}`, //Point to specific ID
     })
-      .done((stories) => {
-        // console.log(stories); //Test
-
-        // Render story to view section.
-        stories.forEach((story) => {
-          const $story = genStoryView(story);
-          $("#story-view").append($story);
-        });
+      .done((story) => {
+         $("#story-view").empty();
+         const $story = genStoryView(story);
+         $("#story-view").append($story);
       })
       .catch((err) => {
         console.log("Error fetching story:", err);
